# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sports-Point API",
    "version": "1.0.0",
    "description": "运动会积分与成绩管理系统对外 API 文档。"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/results": {
      "get": {
        "tags": [
          "成绩"
        ],
        "summary": "分页查询成绩",
        "description": "分页查询成绩记录。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Result"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "page_size": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "页码，从 1 开始",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "description": "每页数量",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "description": "关键词筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department_name",
            "in": "query",
            "required": false,
            "description": "部门名称筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gender",
            "in": "query",
            "required": false,
            "description": "性别筛选：male/female/mixed",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "age_group",
            "in": "query",
            "required": false,
            "description": "组别筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "description": "项目类别筛选：competitive/fun",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scoring_strategy",
            "in": "query",
            "required": false,
            "description": "成绩策略筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_dir",
            "in": "query",
            "required": false,
            "description": "排序方向：asc/desc",
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "tags": [
          "成绩"
        ],
        "summary": "录入成绩",
        "description": "录入个人或团体项目成绩，并按规则计算积分。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResultRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/CreateResultRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "OkResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "ok"
        ],
        "additionalProperties": true
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "参数错误"
          }
        },
        "required": [
          "ok",
          "error"
        ]
      },
...
      "Result": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "成绩 ID"
          },
          "event_name": {
            "type": "string",
            "description": "项目名称"
          },
          "category": {
            "type": "string",
            "description": "项目类别"
          },
          "scoring_strategy": {
            "type": "string",
            "description": "计分策略"
          },
          "age_group": {
            "type": "string",
            "description": "组别"
          },
          "result_type": {
            "type": "string",
            "description": "成绩类型：athlete/team"
          },
          "athlete_type": {
            "type": "string",
            "description": "运动员类型"
          },
          "target_name": {
            "type": "string",
            "description": "运动员名或队伍名"
          },
          "department_name": {
            "type": "string",
            "description": "归属单位"
          },
          "rank": {
            "type": "integer",
            "description": "名次"
          },
          "points": {
            "type": "integer",
            "description": "积分"
          },
          "performance": {
            "type": "string",
            "description": "成绩文本"
          },
          "entered_by": {
            "type": "string",
            "description": "录入人员"
          },
          "created_at": {
            "type": "string",
            "description": "创建时间"
          }
        }
      },
...
      "CreateResultRequest": {
        "type": "object",
        "properties": {
          "event_id": {
            "type": "integer",
            "description": "项目 ID"
          },
          "rank": {
            "type": "integer",
            "description": "名次，不传则自动推定"
          },
          "athlete_type": {
            "type": "string",
            "description": "运动员类型：competitive/fun"
          },
          "athlete_id": {
            "type": "integer",
            "description": "运动员数据库 ID"
          },
          "athlete_no": {
            "type": "string",
            "description": "运动员号（可替代 athlete_id）"
          },
          "team_id": {
            "type": "integer",
            "description": "队伍 ID（团体项目用）"
          },
          "performance": {
            "type": "string",
            "description": "成绩文本"
          },
          "entered_by": {
            "type": "string",
            "description": "录入人员姓名或编号"
          }
        },
        "required": [
          "event_id"
        ]
      },
...
    }
  }
}